+++ R/prcbench.R
@@ -6,13 +6,19 @@
 #'
 #' @section Tool interface:
 #' The \code{\link{create_toolset}} function creates a common interface for
-#'   six different tools that calculate Precision-Recall curves. These tools
+#'   seven different tools that calculate Precision-Recall curves. These tools
 #'   are \href{https://ipa-tys.github.io/ROCR/}{ROCR},
 #'   \href{http://mark.goadrich.com/programs/AUC/}{AUCCalculator},
 #'   \href{https://cran.r-project.org/package=PerfMeas}{PerfMeas},
 #'   \href{https://cran.r-project.org/package=PRROC}{PRROC},
-#'   \href{https://cran.r-project.org/package=precrec}{precrec}, and
-#'   \href{https://yardstick.tidymodels.org/}{yardstick}.
+#'   \href{https://cran.r-project.org/package=precrec}{precrec},
+#'   \href{https://yardstick.tidymodels.org/}{yardstick}, and
+#'   \href{https://scikit-learn.org/stable/modules/generated/sklearn.metrics.precision_recall_curve.html}{scikit-learn}.
+#'
+#' The \code{sklearn} tool is calculated by a standalone Python module that is
+#'   bundled with \code{prcbench} and derived from the scikit-learn source
+#'   code. It requires \code{reticulate}, a working Python installation and
+#'   \code{numpy}, and it is not included in any of the predefined tool sets.
 #'
 #' The \code{\link{create_usrtool}} function helps users to make the same
 #'  interface of the predefined ones for their own tools.